#72 - refactor : Kakao Paging 관련 코드 레포지토리 분리 및 Repository 클래스 삭제
