+++ core/data/src/main/java/com/hegunhee/data/di/RepositoryModule.kt
@@ -1,12 +1,12 @@
 package com.hegunhee.data.di
 
-import com.hegunhee.data.repository.DefaultRepository
+import com.hegunhee.data.repository.kakao.DefaultKakaoPagingRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchGameStreamRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchPagingRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchSearchStreamerRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchStreamRepository
 import com.hegunhee.data.repository.twitch.DefaultTwitchUpdateStreamerRepository
-import com.hegunhee.domain.repository.Repository
+import com.hegunhee.domain.repository.kakao.KakaoPagingRepository
 import com.hegunhee.domain.repository.twitch.TwitchGameStreamRepository
 import com.hegunhee.domain.repository.twitch.TwitchPagingRepository
 import com.hegunhee.domain.repository.twitch.TwitchSearchStreamerRepository
@@ -21,11 +21,7 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 @Module
 abstract class RepositoryModule {
-
-    @Binds
-    @Singleton
-    abstract fun provideRepository(defaultRepository: DefaultRepository): Repository
-
+    
     @Binds
     @Singleton
     abstract fun provideTwitchStreamRepository(defaultTwitchStreamRepository: DefaultTwitchStreamRepository): TwitchStreamRepository
@@ -45,4 +41,8 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun provideTwitchPagingRepository(defaultTwitchPagingRepository: DefaultTwitchPagingRepository): TwitchPagingRepository
+
+    @Binds
+    @Singleton
+    abstract fun provideKakaoPagingRepository(defaultKakaoPagingRepository: DefaultKakaoPagingRepository): KakaoPagingRepository
 }
