+++ core/data/src/test/java/com/hegunhee/data/repository/kakao/SearchPagingDataTest.kt
@@ -3,15 +3,12 @@ package com.hegunhee.data.repository.kakao
 import androidx.paging.PagingData
 import com.hegunhee.data.dataSource.local.LocalDataSource
 import com.hegunhee.data.dataSource.remote.RemoteDataSource
-import com.hegunhee.data.repository.DefaultRepository
 import com.hegunhee.domain.model.kakao.KakaoSearchData
 import com.hegunhee.domain.model.kakao.KakaoSearchSortType
 import com.hegunhee.domain.model.kakao.KakaoSearchType
-import com.hegunhee.domain.model.twitch.SearchData
 import kotlinx.coroutines.flow.count
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.runBlocking
-import org.junit.Assert
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -25,7 +22,7 @@ import org.mockito.kotlin.whenever
 class SearchPagingDataTest {
 
     @InjectMocks
-    private lateinit var sut: DefaultRepository
+    private lateinit var sut: DefaultKakaoPagingRepository
 
     @Mock
     private lateinit var remoteDataSource: RemoteDataSource
@@ -51,7 +48,7 @@ class SearchPagingDataTest {
             ).thenReturn(flowOf(createPagingData(query)))
 
             // When
-            val pagingData = sut.getKakaoSearchPagingData(query, sortType, searchType, size)
+            val pagingData = sut.getSearchPagingData(query, sortType, searchType, size)
 
             // Then
             assertTrue(pagingData.count() == 1)
