+++ core/data/src/test/java/com/hegunhee/data/repository/twitch/SearchDataTest.kt
@@ -4,7 +4,6 @@ import com.hegunhee.data.data.json.twitch.SearchApiData
 import com.hegunhee.data.data.json.twitch.SearchApiDataResponse
 import com.hegunhee.data.dataSource.local.LocalDataSource
 import com.hegunhee.data.dataSource.remote.RemoteDataSource
-import com.hegunhee.data.repository.DefaultRepository
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
