+++ core/domain/src/main/java/com/hegunhee/domain/repository/kakao/KakaoPagingRepository.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.domain.repository
+package com.hegunhee.domain.repository.kakao
 
 import androidx.paging.PagingData
 import com.hegunhee.domain.model.kakao.KakaoSearchData
@@ -6,7 +6,7 @@ import com.hegunhee.domain.model.kakao.KakaoSearchSortType
 import com.hegunhee.domain.model.kakao.KakaoSearchType
 import kotlinx.coroutines.flow.Flow
 
-interface Repository {
+interface KakaoPagingRepository {
 
-    suspend fun getKakaoSearchPagingData(query: String, sortType: KakaoSearchSortType, searchType: KakaoSearchType, size: Int) : Flow<PagingData<KakaoSearchData>>
+    suspend fun getSearchPagingData(query: String, sortType: KakaoSearchSortType, searchType: KakaoSearchType, size: Int) : Flow<PagingData<KakaoSearchData>>
 }
